requirements.txt
python-telegram-bot==21.3
httpx==0.28.1
pydantic==2.7.1
pyyaml==6.0.1
python-dotenv==1.0.0
pdfminer.six==20231228
python-docx==1.1.2
python-magic==0.4.27
fastapi==0.110.0
spoon-ai-sdk==0.3.3
spoon-toolkits==0.2.2